requirements.txt
Django~=2.0.6
djangorestframework
markdown
django-filter
pycep-correios
